Refactor auth and api service
