--- api_service/api_base.ts
@@ -1,2 +1,0 @@
-const localIpAddress = "192.168.8.147";
-export const baseApiUrl = `http://${localIpAddress}:8000`;
